Fallback to Gray when no OriginalForegroundColor
While using the NuGet Console in Visual Studio I got errors when listing directory contents. It was because somehow the $Host.UI.RawUI.ForegroundColor was = -1. I don't know how often this situation arises, but the fix above makes everything work on my system. Essentially we're using some .NET code to ensure the value in the ForegroundColor property is a valid ConsoleColor. If not we simply use "Gray".